Add files via upload
Introduction
How stable is the relationship between unemployment rate and the inflation rate for the past decades in the United States? The relation between employment and inflation is called Phillips Curve. Many students in Macroeconomics have been taught that unemployment and inflation are negatively correlated. If this relationship is steady or even existing, there would be an incentive for policymakers to forgo inflation in order to achieve lower unemployment. In 1960, Paul Samuelson and Robert Solow applied the Phillips Curve study to The United States, at that time the relation appeared to be solid. Considering the results, they proposed that the price level should increase between 4% and 5% in a year. For the unemployment rate to be not more than 3%, this estimation was subject to a moderate GDP growth. 
What motivates me to investigate this issue is the desire to understand under what economic circumstances is the relationship between inflation and unemployment manipulable by policy makers. Is full employment possible at a given economic setting? If yes, could this conditions be quantitatively determined? This work probably does not clarify these doubts. Yet, I hope to objectively determine how statistically significant the relationship between unemployment and inflation is for the sample chosen, with the aids of econometric analysis.
 Particularly, the concern is the statically relationship between The inflation rate and the unemployment rate for United States sample in the period 1970:01-2015:12. Does it provide empirical evidence for the Phillips curve? Applying appropriate statistical techniques. I will test the following hypothesis: the unemployment rate is negatively correlated with the rate of inflation.